+++ ums_backend/src/main/java/repository/MentorRepository.java
@@ -3,6 +3,6 @@ package repository;
 import models.mentors.Mentor;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface MentorRepository extends JpaRepository<Mentor, Long>
+public interface MentorRepository extends JpaRepository<Mentor, Integer>
 {
 }
